Tweak entries drag and drop
If you were to begin to drag an entry around, it was unclear what would
happen if you were to drag the entry back in to place. Now it does the
expected behavior which is to turn off the border on any other entry
that might have the border and dropping the entry back in to place
changes nothing.