Set up timer controller in app didFinishLaunching
The app delegate’s `timerController` was getting set in the
PersistenceController callback which is normally fine but because
extensions (watch glance, notification actions) rely on the
timerController in their callbacks, waiting to initialize it was
causing problems. So now it gets initialized straight away so that it
is ready for use as soon as the app is awoken.